Move TuitionFeesActivity content to TuitionFeesFragment
